Conflicts:
	kernel/core/Scene.cpp

+++ kernel/core/Receiver.cpp
@@ -31,7 +31,7 @@ using namespace Eigen;
 using namespace std;
 namespace OpenPSTD {
     namespace Kernel {
-        Receiver::Receiver(vector<float> location, shared_ptr<PSTDSettings> config, int id,
+        Receiver::Receiver(vector<float> location, shared_ptr<PSTDSettings> config, unsigned long id,
                            shared_ptr<Domain> container) : x(location.at(0)), y(location.at(1)), z(location.at(2)) {
             this->config = config;
             this->location = location;
